Updated post-processing and metadata file
The metadata file contains the right tracer name now and have a default
number of events before post-processing. The name of the field
containing the number of events also has been updated.
Updated the way the offsets are processed in post-processing, to be more
precise.
Renamed the directory containing the CTF trace. Before, it was always
named CTF_trace and was not convenient to open several synchronized
traces on tools such as TraceCompass. Now the trace is named
rocprof_ctf_trace before the post-processing step. After the
post-processing step, the pid, date and hour are appened to the name.