Add script to regenerate barectf files and improve post-processing
Added a script to regenerate barectf files when using --barectf-generate
option when building the plugin.
Added a field containing the ROCm version installed on the machine. This
field is computed when building the plugin if using --barectf-generate
command. It is also provided by default on the metadata file present in
the directory.
Changed the way the offsets and number of events are written in the
metadata file at post-processing to make it more robust to metadata file
changes.